+++ Organization.md
@@ -1,0 +1,12 @@
+
+# Organization of code
+
+tiffImage.c & .h
+ - image and file struct
+ - image checking code
+
+tiffReader.c
+ - read image and turn it to image and file struct data
+
+tiffTagStorage.c & .h
+ - data tag
