+++ tiffTagStorage.c
@@ -7,9 +7,8 @@ tiffDataTag_t dataTag(WORD tagId, WORD dataType, size_t dataCount) {
     tiffDataTag_t tmp;
     tmp.tagId = tagId;
     tmp.dataType = dataType;
-    tmp.dataTypeSize = sizeOfTypeLookupTable[dataType];
     tmp.dataCount = dataCount;
-    tmp.data = malloc(tmp.dataTypeSize * dataCount);
+    tmp.data = malloc(getTypeSizeOf(dataType) * dataCount);
     return tmp;
 }
 
@@ -17,9 +16,9 @@ void freeDataTag(tiffDataTag_t t) {
     free(t.data);
 }
 
-int main(void) {
-    printf("test\n");
-    tiffDataTag_t dt = dataTag(1, 1, 1);
+// int main(void) {
+//     printf("test\n");
+//     tiffDataTag_t dt = dataTag(1, 1, 1);
 
-    freeDataTag(dt);
-}
+//     freeDataTag(dt);
+// }
